+++ .yarn/versions/cbe1810e.yml
@@ -1,0 +1,2 @@
+releases:
+  "@iguhealth/server": patch
